Simplify request handling'

+++ src/main/kotlin/io/swagger/petstore/infrastructure/ApiClient.kt
@@ -8,12 +8,14 @@ open class ApiClient(val baseUrl: String) {
         protected val ContentType = "Content-Type"
         protected val Accept = "Accept"
         protected val JsonMediaType = "application/json"
+        protected val FormDataMediaType = "multipart/form-data"
+        protected val XmlMediaType = "application/xml"
 
         @JvmStatic
         val client : OkHttpClient = OkHttpClient()
 
         @JvmStatic
-        var defaultHeaders: Map<String, String> by ApplicationDelegates.setOnce(mapOf())
+        var defaultHeaders: Map<String, String> by ApplicationDelegates.setOnce(mapOf(ContentType to JsonMediaType, Accept to JsonMediaType))
 
         @JvmStatic
         val jsonHeaders: Map<String, String> = mapOf(ContentType to JsonMediaType, Accept to JsonMediaType)
@@ -24,17 +26,35 @@ open class ApiClient(val baseUrl: String) {
             return RequestBody.create(
                     MediaType.parse(mediaType), content
             )
-        } else if(mediaType == "application/json") {
+        } else if(mediaType == FormDataMediaType) {
+            var builder = FormBody.Builder()
+            // content's type *must* be Map<String, Any>
+            @Suppress("UNCHECKED_CAST")
+            (content as Map<String,String>).forEach { key, value ->
+                builder = builder.add(key, value)
+            }
+            return builder.build()
+        }  else if(mediaType == JsonMediaType) {
             return RequestBody.create(
                     MediaType.parse(mediaType), Serializer.moshi.adapter(T::class.java).toJson(content)
             )
+        } else if (mediaType == XmlMediaType) {
+            TODO("xml not currently supported.")
         }
 
         // TODO: this should be extended with other serializers
         TODO("requestBody currently only supports JSON body and File body.")
     }
 
-    protected fun request(requestConfig: RequestConfig, body : Any? = null): Response {
+    inline protected fun <reified T: Any?> responseBody(body: ResponseBody?, mediaType: String = JsonMediaType): T? {
+        if(body == null) return null
+        return when(mediaType) {
+            JsonMediaType -> Serializer.moshi.adapter(T::class.java).fromJson(body.source())
+            else -> TODO()
+        }
+    }
+
+    inline protected fun <reified T: Any?> request(requestConfig: RequestConfig, body : Any? = null): ApiResponse<T?> {
         val httpUrl = HttpUrl.parse(baseUrl) ?: throw IllegalStateException("baseUrl is invalid.")
 
         var urlBuilder = httpUrl.newBuilder()
@@ -45,13 +65,24 @@ open class ApiClient(val baseUrl: String) {
         val url = urlBuilder.build()
         val headers = requestConfig.headers + defaultHeaders
 
+        if(headers[ContentType] ?: "" == "") {
+            throw kotlin.IllegalStateException("Missing Content-Type header. This is required.")
+        }
+
+        if(headers[Accept] ?: "" == "") {
+            throw kotlin.IllegalStateException("Missing Accept header. This is required.")
+        }
+
+        val contentType = (headers[ContentType] as String).substringBefore(";").toLowerCase()
+        val accept = (headers[Accept] as String).substringBefore(";").toLowerCase()
+
         var request : Request.Builder =  when (requestConfig.method) {
             RequestMethod.DELETE -> Request.Builder().url(url).delete()
             RequestMethod.GET -> Request.Builder().url(url)
             RequestMethod.HEAD -> Request.Builder().url(url).head()
-            RequestMethod.PATCH -> Request.Builder().url(url).patch(requestBody(body!!, headers.getOrDefault(ContentType, JsonMediaType)))
-            RequestMethod.PUT -> Request.Builder().url(url).put(requestBody(body!!, headers.getOrDefault(ContentType, JsonMediaType)))
-            RequestMethod.POST -> Request.Builder().url(url).post(requestBody(body!!, headers.getOrDefault(ContentType, JsonMediaType)))
+            RequestMethod.PATCH -> Request.Builder().url(url).patch(requestBody(body!!, contentType))
+            RequestMethod.PUT -> Request.Builder().url(url).put(requestBody(body!!, contentType))
+            RequestMethod.POST -> Request.Builder().url(url).post(requestBody(body!!, contentType))
             RequestMethod.OPTIONS -> Request.Builder().url(url).method("OPTIONS", null)
         }
 
@@ -60,12 +91,7 @@ open class ApiClient(val baseUrl: String) {
         val realRequest = request.build()
         val response = client.newCall(realRequest).execute()
 
-        return response
-    }
-
-    protected inline fun <reified T : Any, reified K : Any?> jsonWithBody(requestConfig: RequestConfig, body : K? = null): ApiResponse<T?> {
-        val response = request(requestConfig.copy(headers = requestConfig.headers + jsonHeaders), body)
-
+        // TODO: handle specific mapping types. e.g. Map<int, Class<?>>
         when {
             response.isRedirect -> return Redirection(
                     response.code(),
@@ -77,7 +103,7 @@ open class ApiClient(val baseUrl: String) {
                     response.headers().toMultimap()
             )
             response.isSuccessful -> return Success(
-                    Serializer.moshi.adapter(T::class.java).fromJson(response.body()?.string()),
+                    responseBody(response.body(), accept),
                     response.code(),
                     response.headers().toMultimap()
             )
@@ -94,6 +120,4 @@ open class ApiClient(val baseUrl: String) {
             )
         }
     }
-
-    protected inline fun <reified T : Any> json(requestConfig: RequestConfig): ApiResponse<T?> = jsonWithBody(requestConfig, null as? Any?)
 }
